ajout du contrôle de la vue édition des utilisateurs par les rôles.

+++ app/Http/Controllers/Admin/UsersController.php
@@ -6,6 +6,7 @@ use App\Http\Controllers\Controller;
 use App\User;
 use App\Role;
 use Illuminate\Http\Request;
+use Illuminate\support\Facades\Gate;
 
 class UsersController extends Controller
 {
@@ -65,6 +66,10 @@ class UsersController extends Controller
      */
     public function edit(User $user)
     {
+        if (Gate::denies('edit-users')) {
+            return redirect()->route('admin.users.index');
+        }
+
         $roles = Role::all();
 
         return view('admin.users.edit', [
@@ -95,6 +100,10 @@ class UsersController extends Controller
      */
     public function destroy(User $user)
     {
+        if (Gate::denies('delete-users')) {
+            return redirect()->route('admin.users.index');
+        }
+
         $user->roles()->detach();
         $user->delete();
